refactor: split GenreGuide, move domain data, fix audit issues
Component extraction:
- ExposureMatrix.tsx (72 lines) — rule-of-500 grid
- shared/MarkPips.tsx (49 lines) — MarkPips, PickStar, FieldVal
- GenreGuide.tsx reduced from 838 to 648 lines

Domain data moved to src/data/genres.ts:
- GENRE_EQUIPMENT, GENRE_DEFAULTS, ASTRO_ISO_BY_EV
- FL_CHIPS, FL_RANGES, ND_OPTIONS
- MATRIX_FL_COLS, MATRIX_APERTURES

Bug fixes:
- Dead code: removed `false ? styles.cellRight` leftover
- WR sort key: was "brand", now non-sortable
- FieldVal: CSS classes instead of inline styles
- Renamed aoV→selectedFl, mp→sensorMp, crop→cropFactor
- Removed empty CSS rules and duplicate comments

188 tests passing.
