+++ src/components/interactive/GenreGuide/GenreGuide.module.css
@@ -290,10 +290,6 @@
 }
 
 /* ==========================================================================
-   Results
-   ========================================================================== */
-
-/* ==========================================================================
    Filter panel — above lens table
    ========================================================================== */
 
@@ -340,9 +336,7 @@
   padding: 10px 14px;
 }
 
-.matrix {
-  /* no extra spacing — inside matrixPanel */
-}
+/* .matrix has no extra spacing — sits inside matrixPanel */
 
 .matrixTitle {
   font-size: var(--font-size-sm);
@@ -472,6 +466,23 @@
 
 .markDash {
   color: #8a8070;
+}
+
+/* Field value colors — traffic-light scale for optical scores (0-2) */
+.fieldViable {
+  color: #6aaa70;  /* green — good optical performance */
+}
+
+.fieldMarginal {
+  color: #b09040;  /* amber — acceptable */
+}
+
+.fieldOver {
+  color: #905030;  /* red — poor */
+}
+
+.fieldMono {
+  font-family: var(--font-mono);
 }
 
 .topStar {
