+++ src/data/genres.ts
@@ -234,7 +234,134 @@ const genreSceneFilter: Record<ScoredGenre, (ev: number) => boolean> = {
 };
 
 // =============================================================================
+// ND FILTER OPTIONS
+// =============================================================================
+
+const ND_OPTIONS = [
+  { label: "ND2", factor: 2 },
+  { label: "ND4", factor: 4 },
+  { label: "ND8", factor: 8 },
+  { label: "ND64", factor: 64 },
+  { label: "ND1000", factor: 1000 },
+];
+
+// =============================================================================
+// FL CHIPS — per-genre focal-length presets
+// =============================================================================
+
+const FL_CHIPS: Record<string, { label: string; fl: number }[]> = {
+  default: [
+    { label: "Ultra-wide", fl: 12 },
+    { label: "Wide", fl: 24 },
+    { label: "Standard", fl: 50 },
+    { label: "Tele", fl: 135 },
+    { label: "Super-tele", fl: 300 },
+  ],
+  portrait: [
+    { label: "Group", fl: 15 },
+    { label: "Indoor", fl: 33 },
+    { label: "Outdoor", fl: 57 },
+  ],
+  sportWildlife: [
+    { label: "Standard", fl: 50 },
+    { label: "Tele", fl: 135 },
+    { label: "Super-tele", fl: 300 },
+  ],
+  macro: [
+    { label: "Standard", fl: 50 },
+    { label: "Tele", fl: 90 },
+    { label: "Long macro", fl: 200 },
+  ],
+};
+
+// =============================================================================
+// FL RANGES — lens must overlap the range to show (actual X-mount focal lengths)
+// =============================================================================
+
+// Ultra-wide: 6-15mm, Wide: 16-27mm, Standard: 28-56mm, Tele: 57-150mm, Super-tele: 151+
+const FL_RANGES: Record<number, [number, number]> = {
+  12:  [6, 15],      // Ultra-wide
+  15:  [12, 24],     // Portrait group
+  24:  [16, 27],     // Wide
+  33:  [24, 45],     // Portrait indoor
+  50:  [28, 56],     // Standard
+  57:  [40, 75],     // Portrait outdoor
+  90:  [57, 150],    // Tele / macro tele
+  135: [57, 150],    // Tele
+  200: [100, 300],   // Long macro
+  300: [151, 600],   // Super-tele
+};
+
+// =============================================================================
+// GENRE DEFAULTS — initial EV, ISO, FL per genre
+// =============================================================================
+
+const GENRE_DEFAULTS: Record<ScoredGenre, { ev: number; iso: number; fl: number }> = {
+  astro:        { ev: -7, iso: 3200, fl: 12 },
+  landscape:    { ev: 9,  iso: 100,  fl: 24 },
+  architecture: { ev: 7,  iso: 200,  fl: 12 },
+  street:       { ev: 2,  iso: 6400, fl: 24 },
+  travel:       { ev: 11, iso: 400,  fl: 24 },
+  portrait:     { ev: 10, iso: 200,  fl: 57 },
+  sport:        { ev: 13, iso: 800,  fl: 135 },
+  wildlife:     { ev: 9,  iso: 3200, fl: 300 },
+  macro:        { ev: 10, iso: 200,  fl: 90 },
+};
+
+// =============================================================================
+// GENRE EQUIPMENT — recommended gear per genre
+// =============================================================================
+
+const GENRE_EQUIPMENT: Record<string, string[]> = {
+  astro:        ["Star tracker", "Sturdy tripod", "Lens heater", "Dew shield", "Light pollution filter", "External power bank", "Remote intervalometer", "Bahtinov mask", "Spare batteries", "Red light headlamp"],
+  landscape:    ["Sturdy tripod", "Ball head", "L-bracket", "Remote shutter", "Filter set (CPL & ND)", "Graduated ND filter", "Rain cover", "Spare batteries", "Fast memory cards", "Outdoor camera backpack"],
+  architecture: ["Sturdy tripod", "Geared tripod head", "L-bracket", "Remote shutter", "Tethering cable", "Field monitor", "Filter set (CPL & Graduated ND)", "Anti-reflective lens cover", "Color checker", "Power bank"],
+  street:       ["Thumb grip", "Compact bag", "Quick-adjust sling strap", "Mini travel tripod", "Flash (Compact)", "Filter set (ND, CPL & Black Mist)", "Fast memory cards", "Spare batteries", "Soft shutter release button", "Rain cover"],
+  travel:       ["Travel tripod", "Adjustable sling strap", "Travel camera backpack", "L-bracket", "Filter set (ND & Polarizer)", "Anti-reflection hood", "Remote shutter", "Fast memory cards", "Spare batteries", "Portable SSD"],
+  portrait:     ["Tripod", "Speedlight", "Flash trigger", "Light stand", "Constant light", "Light modifiers", "Reflector", "Tethering cable", "Color checker", "Remote shutter"],
+  sport:        ["Monopod", "Battery grip", "Dual harness", "Teleconverter", "Remote trigger", "Fast memory cards", "Spare batteries", "Rain cover", "Power bank", "Camera gear bag"],
+  wildlife:     ["Gimbal head", "Tripod/Monopod", "Teleconverter", "Beanbag", "Shoulder sling strap", "Lens covers", "Fast memory cards", "Spare batteries", "Power bank", "Wildlife backpack"],
+  macro:        ["Sturdy tripod", "Macro rail", "Ring flash", "Diffuser", "Extension tubes", "Remote shutter", "Reflector", "Focus stacking software", "Spare batteries", "Macro shooting tent"],
+};
+
+// =============================================================================
+// ASTRO ISO BY EV — suggested ISO per EV for astro, balanced at f/2.8
+// =============================================================================
+
+const ASTRO_ISO_BY_EV: Record<number, number> = {
+  1: 100, 0: 200, [-1]: 200, [-2]: 400, [-3]: 400,
+  [-4]: 800, [-5]: 1600, [-6]: 1600, [-7]: 3200, [-8]: 6400,
+};
+
+// =============================================================================
+// EXPOSURE MATRIX CONSTANTS — focal lengths and apertures for astro grid
+// =============================================================================
+
+const MATRIX_FL_COLS: Record<number, number[]> = {
+  12:  [8, 10, 12, 14],
+  24:  [16, 18, 23, 27],
+  50:  [33, 35, 50, 56],
+  135: [60, 80, 90, 100, 135],
+  300: [150, 200, 300, 400],
+};
+
+const MATRIX_APERTURES = [1.0, 1.4, 2.0, 2.8, 4.0, 5.6, 8, 11];
+
+// =============================================================================
 // EXPORTS
 // =============================================================================
 
-export { genreConfigs, evScenes, genreEvLabels, genreSceneFilter };
+export {
+  genreConfigs,
+  evScenes,
+  genreEvLabels,
+  genreSceneFilter,
+  ND_OPTIONS,
+  FL_CHIPS,
+  FL_RANGES,
+  GENRE_DEFAULTS,
+  GENRE_EQUIPMENT,
+  ASTRO_ISO_BY_EV,
+  MATRIX_FL_COLS,
+  MATRIX_APERTURES,
+};
